refactor AppRouter as a component
